Split dbConnect from db.js to new file db-connect.js so that tests will work
